requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
DateTime==4.7
discord.py @ git+https://github.com/Rapptz/discord.py@24bdb44d54686448a336ea6d72b1bf8600ef7220
frozenlist==1.3.3
greenlet==2.0.1
idna==3.4
lxml==4.9.2
multidict==6.0.2
pytz==2022.6
requests==2.28.1
requests-toolbelt==0.10.1
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
urllib3==1.26.13
yarl==1.8.1
